copy some static files add discriminator values

+++ raml-generic-generator/src/main/java/io/vrap/rmf/raml/generic/generator/php/PhpGenerator.java
@@ -1,11 +1,14 @@
 package io.vrap.rmf.raml.generic.generator.php;
 
+import com.google.common.collect.Lists;
+import com.google.common.io.Resources;
 import io.vrap.rmf.raml.generic.generator.Generator;
 import io.vrap.rmf.raml.model.modules.Api;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.util.List;
 
 public class PhpGenerator implements Generator {
 
@@ -14,7 +17,9 @@ public class PhpGenerator implements Generator {
             File[] files = outputPath.listFiles();
             if (files != null) {
                 for (File file : files) {
-                    Files.deleteIfExists(file.toPath());
+                    if (file.isFile()) {
+                        Files.deleteIfExists(file.toPath());
+                    }
                 }
             }
         } else {
@@ -22,6 +27,18 @@ public class PhpGenerator implements Generator {
         }
 
         TypesGenerator generator = new TypesGenerator();
-        generator.generate(api.getTypes(), outputPath);
+        generator.generate(api.getTypes(), new File(outputPath, "target") );
+
+        List<File> files = Lists.newArrayList(
+            new File(Resources.getResource("./templates/php/composer.json").getFile())
+        );
+
+        files.forEach(file -> {
+            try {
+                Files.copy(file.toPath(), new File(outputPath, file.getName()).toPath());
+            } catch (IOException e) {
+                System.out.println(e);
+            }
+        });
     }
 }
